Modified files in the search algorithm to use double instead of float. Removed some trailing tabs/whitespaces.

+++ beameff2_CPP/ScanDataRaster.h
@@ -130,7 +130,7 @@ public:
     bool unwrapPhase();
     ///< unwraps the phase array
 
-    float calcPhaseEfficiency(float p[], float azNominal, float elNominal,
+    double calcPhaseEfficiency(double p[], float azNominal, float elNominal,
                               bool approx = false, bool reduceSubreflector = false) const;
     ///< calculate phase efficiency for the given phase center model p[].
     ///< called from within FitPhase()
@@ -140,7 +140,7 @@ public:
     ///< @param reduceSubreflector:  if true, reduce the subreflector radius to calculate over.
     ///< @return phase efficiency in 0-1.
 
-    float calcAmplitudeEfficiency(float p[], float azActual, float elActual) const;
+    double calcAmplitudeEfficiency(double p[], float azActual, float elActual) const;
     ///< calculate amplitude efficiency at the given amplitude fit model p[].
     ///< called from within FitAmplitude()
     ///< @param p: array giving amplitude fit model
